chores: assign weave_knowledge_graph.py as main script, original script has been moved to weave_knowledge_graph_original.py
